pyproject.toml
[project]
name = "trend-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "imbalanced-learn>=0.14.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "tensorflow>=2.16.1",
    "mlflow>=3.8.0",
    "dvc>=3.64.2",
    "mlforecast>=1.0.2",
    "ccxt>=4.5.29",
    "metatrader5>=5.0.5430",
    "xgboost>=3.1.2",
    "lightgbm>=4.6.0",
    "neuralforecast==3.1.2",
    "apscheduler>=3.10.4",
]

[project.scripts]
# Remove previous script as we shift to MLflow serving

[tool.setuptools.packages.find]
where = ["."]
include = ["trend_analysis*"]
exclude = ["models*", "data*", "notebooks*"]
